Use double-precision floating point

+++ ramon_calibration/include/accel_calibration.h
@@ -24,38 +24,38 @@ namespace ramon_calibration
 	{
 		public:
 			AccelCalibration();
-			AccelCalibration(Eigen::Vector3f noise);
-			AccelCalibration(Eigen::Matrix3f T_init, Eigen::Matrix3f K_init, Eigen::Vector3f b_init);
-			AccelCalibration(Eigen::Matrix3f T_init, Eigen::Matrix3f K_init, Eigen::Vector3f b_init, Eigen::Vector3f noise);
+			AccelCalibration(Eigen::Vector3d noise);
+			AccelCalibration(Eigen::Matrix3d T_init, Eigen::Matrix3d K_init, Eigen::Vector3d b_init);
+			AccelCalibration(Eigen::Matrix3d T_init, Eigen::Matrix3d K_init, Eigen::Vector3d b_init, Eigen::Vector3d noise);
 			~AccelCalibration();
-			Eigen::LevenbergMarquardtSpace::Status calibrate(const Eigen::Ref<const Eigen::MatrixXf> accel_values);
-			int correctValues(const Eigen::Ref<const Eigen::MatrixX3f> accel_values, Eigen::Ref<Eigen::MatrixX3f> accel_values_corr);
-			int getBias(Eigen::Ref<Eigen::Vector3f> b);
-			int getParams(Eigen::Ref<Eigen::Matrix3f> T, Eigen::Ref<Eigen::Matrix3f> K, Eigen::Ref<Eigen::Vector3f> b);
-			int getT(Eigen::Ref<Eigen::Matrix3f> T);
-			int getK(Eigen::Ref<Eigen::Matrix3f> K);
+			Eigen::LevenbergMarquardtSpace::Status calibrate(const Eigen::Ref<const Eigen::MatrixXd> accel_values);
+			int correctValues(const Eigen::Ref<const Eigen::MatrixX3d> accel_values, Eigen::Ref<Eigen::MatrixX3d> accel_values_corr);
+			int getBias(Eigen::Ref<Eigen::Vector3d> b);
+			int getParams(Eigen::Ref<Eigen::Matrix3d> T, Eigen::Ref<Eigen::Matrix3d> K, Eigen::Ref<Eigen::Vector3d> b);
+			int getT(Eigen::Ref<Eigen::Matrix3d> T);
+			int getK(Eigen::Ref<Eigen::Matrix3d> K);
 			
 			int printCalibratedValues(void);
 			
-			void setInitalValues(Eigen::Matrix3f T_init, Eigen::Matrix3f K_init, Eigen::Vector3f b_init);
-			void setNoise(Eigen::Vector3f noise);
+			void setInitalValues(Eigen::Matrix3d T_init, Eigen::Matrix3d K_init, Eigen::Vector3d b_init);
+			void setNoise(Eigen::Vector3d noise);
 		private:
 			int n_;
 			int m_;
 			bool calibrated_;
 			
-			Eigen::Matrix3f T_;
-			float alpha_yz_, alpha_zy_, alpha_zx_;
+			Eigen::Matrix3d T_;
+			double alpha_yz_, alpha_zy_, alpha_zx_;
 
-			Eigen::Matrix3f K_;
-			float s_x_, s_y_, s_z_;
+			Eigen::Matrix3d K_;
+			double s_x_, s_y_, s_z_;
 
-			Eigen::Vector3f b_;
-			float b_x_, b_y_, b_z_;
+			Eigen::Vector3d b_;
+			double b_x_, b_y_, b_z_;
 
 			bool use_noise_;
-			Eigen::Vector3f noise_;
-			float noise_x_, noise_y_, noise_z_;
+			Eigen::Vector3d noise_;
+			double noise_x_, noise_y_, noise_z_;
 	};
 };
